+++ json_serializable/test/src/to_from_json_test_input.dart
@@ -12,6 +12,8 @@ dynamic _toDynamic(dynamic input) => null;
 
 Object _toObject(Object input) => null;
 
+String _toStringFromObject(Object input) => null;
+
 @ShouldThrow(
   'Error with `@JsonKey` on `field`. The `fromJson` function `_toInt` '
   'return type `int` is not compatible with field type `String`.',
@@ -25,7 +27,7 @@ class BadFromFuncReturnType {
 
 @ShouldThrow(
   'Error with `@JsonKey` on `field`. The `fromJson` function '
-  '`_twoArgFunction` must have one positional paramater.',
+  '`_twoArgFunction` must have one positional parameter.',
   element: 'field',
 )
 @JsonSerializable()
@@ -71,7 +73,7 @@ class BadToFuncReturnType {
 
 @ShouldThrow(
   'Error with `@JsonKey` on `field`. The `toJson` function '
-  '`_twoArgFunction` must have one positional paramater.',
+  '`_twoArgFunction` must have one positional parameter.',
   element: 'field',
 )
 @JsonSerializable()
@@ -81,12 +83,12 @@ class InvalidToFunc2Args {
 }
 
 @ShouldGenerate(
-  "_toObject(json['field'])",
+  "_toStringFromObject(json['field'])",
   contains: true,
 )
 @JsonSerializable()
 class ObjectConvertMethods {
-  @JsonKey(fromJson: _toObject, toJson: _toObject)
+  @JsonKey(fromJson: _toStringFromObject, toJson: _toObject)
   String field;
 }
 
@@ -175,7 +177,7 @@ String _noArgs() => null;
 
 @ShouldThrow(
   'Error with `@JsonKey` on `field`. The `fromJson` function '
-  '`_noArgs` must have one positional paramater.',
+  '`_noArgs` must have one positional parameter.',
   element: 'field',
 )
 @JsonSerializable(createToJson: false)
@@ -188,7 +190,7 @@ String _twoArgs(a, b) => null;
 
 @ShouldThrow(
   'Error with `@JsonKey` on `field`. The `fromJson` function '
-  '`_twoArgs` must have one positional paramater.',
+  '`_twoArgs` must have one positional parameter.',
   element: 'field',
 )
 @JsonSerializable(createToJson: false)
@@ -201,7 +203,7 @@ String _oneNamed({a}) => null;
 
 @ShouldThrow(
   'Error with `@JsonKey` on `field`. The `fromJson` function '
-  '`_oneNamed` must have one positional paramater.',
+  '`_oneNamed` must have one positional parameter.',
   element: 'field',
 )
 @JsonSerializable(createToJson: false)